+++ karaoke-app/src/main/kotlin/com/svoemesto/karaokeapp/model/Uuids.kt
@@ -6,6 +6,8 @@ import java.io.Serializable
 import java.sql.ResultSet
 import java.sql.SQLException
 import java.sql.Statement
+import java.sql.Timestamp
+import java.time.Instant
 
 class Uuids(var id: Int, var uuid: String, val database: KaraokeConnection = WORKING_DATABASE) : Serializable {
 //    var id: Int = 0
@@ -13,6 +15,10 @@ class Uuids(var id: Int, var uuid: String, val database: KaraokeConnection = WOR
     fun save() {
 
         val connection = database.getConnection()
+        if (connection == null) {
+            println("[${Timestamp.from(Instant.now())}] Невозможно установить соединение с базой данных")
+            return
+        }
         val sql = "UPDATE tbl_uuids SET " +
                 "uuid = ?, " +
                 "WHERE id = ?"
@@ -51,6 +57,10 @@ class Uuids(var id: Int, var uuid: String, val database: KaraokeConnection = WOR
             val sql = uuid.getSqlToInsert()
 
             val connection = database.getConnection()
+            if (connection == null) {
+                println("[${Timestamp.from(Instant.now())}] Невозможно установить соединение с базой данных")
+                return null
+            }
             val ps = connection.prepareStatement(sql)
             ps.executeUpdate(sql, Statement.RETURN_GENERATED_KEYS ) //, Statement.RETURN_GENERATED_KEYS)
 //            val rs = ps.generatedKeys
@@ -69,6 +79,10 @@ class Uuids(var id: Int, var uuid: String, val database: KaraokeConnection = WOR
         fun loadList(args: Map<String, String> = emptyMap(), database: KaraokeConnection): List<Uuids> {
 
             val connection = database.getConnection()
+            if (connection == null) {
+                println("[${Timestamp.from(Instant.now())}] Невозможно установить соединение с базой данных")
+                return emptyList()
+            }
             var statement: Statement? = null
             var rs: ResultSet? = null
             var sql: String
@@ -107,6 +121,10 @@ class Uuids(var id: Int, var uuid: String, val database: KaraokeConnection = WOR
         fun delete(id: Int, database: KaraokeConnection) {
 
             val connection = database.getConnection()
+            if (connection == null) {
+                println("[${Timestamp.from(Instant.now())}] Невозможно установить соединение с базой данных")
+                return
+            }
             val sql = "DELETE FROM tbl_uuids WHERE id = ?"
             val ps = connection.prepareStatement(sql)
             var index = 1
